got textures to render, removed all previous rendering code

+++ src/main/java/com/dessertion/icssummative/game/state/MainGameState.java
@@ -14,7 +14,7 @@ public class MainGameState implements State{
 	}
 	
 	@Override
-	public void update(float interval) {
+	public void update(double interval) {
 	
 	}
